+++ ChangeLog.md
@@ -3,6 +3,7 @@ Redis Sessions change log
 
 ## ?.?.? / ????-??-??
 
+* Implement session timeout properly - @thekid
 * Fixed error message when authentication fails - @thekid
 
 ## 0.1.0 / 2019-08-23